# swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This is the API documentation for Book-A-Meal web app.",
    "version" : "1.0.0",
    "title" : "Book-A-Meal API Documentation",
    "contact" : {
      "email" : "onnassiz@gmail.com"
    },
    "license" : {
      "name" : "Book-A-Meak 1.0",
      "url" : "https://mealbooker.herokuapp.com/api/v1"
    }
  },
  "host" : "mealbooker.herokuapp.com",
  "basePath" : "/api/v1",
...
  "schemes" : [ "https" ],
  "paths" : {
...
    "/order/{id}" : {
      "get" : {
        "tags" : [ "Order" ],
        "summary" : "Returns a single order object with the meals in that order",
        "description" : "",
        "operationId" : "getOrderById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "api_key",
          "in" : "header",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "id",
          "in" : "path",
          "description" : "Id of order to return",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Order"
            }
          },
          "403" : {
            "description" : "Forbidden",
            "schema" : {
              "$ref" : "#/definitions/ResponseMessage"
            }
          },
          "404" : {
            "description" : "Order not found",
            "schema" : {
              "$ref" : "#/definitions/ResponseMessage"
            }
          }
        },
        "security" : [ {
          "book_meal_auth" : [ "read:orders" ]
        } ]
      },
      "put" : {
        "tags" : [ "Order" ],
        "summary" : "Update an instance of a order and its associating meals",
        "description" : "",
        "operationId" : "updateOrder",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "A order model",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Menu"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "description" : "Id of order to update",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Order successfully updated",
            "schema" : {
              "$ref" : "#/definitions/Order"
            }
          },
          "400" : {
            "description" : "Invalid Input",
            "schema" : {
              "$ref" : "#/definitions/ResponseMessage"
            }
          },
          "403" : {
            "description" : "Forbidden",
            "schema" : {
              "$ref" : "#/definitions/ResponseMessage"
            }
          },
          "404" : {
            "description" : "Order not found",
            "schema" : {
              "$ref" : "#/definitions/ResponseMessage"
            }
          }
        },
        "security" : [ {
          "book_meal_auth" : [ "write:menus", "read:menus" ]
        } ]
      }
    },
...
  },
  "securityDefinitions" : {
    "book_meal_auth" : {
      "type" : "oauth2",
      "authorizationUrl" : "https://mealbooker.herokuapp.com/api/v1/auth",
      "flow" : "implicit",
      "scopes" : {
        "write:meals" : "create/modify meals in database",
        "read:meals" : "read meals",
        "write:menus" : "create/modify menus in database",
        "read:menus" : "read menus",
        "write:orders" : "create/modify orders in database",
        "read:orders" : "read orders",
        "write:profiles" : "create/modify profile in database",
        "read:profiles" : "read profile"
      }
    },
    "api_key" : {
      "type" : "apiKey",
      "name" : "api_key",
      "in" : "header"
    }
  },
  "definitions" : {
    "Order" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "meals" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Meal"
          }
        }
      },
      "example" : {
        "id" : "id",
        "meals" : [ {
          "price" : 0,
          "imageUrl" : "imageUrl",
          "name" : "name",
          "description" : "description",
          "id" : "id",
          "category" : "category"
        }, {
          "price" : 0,
          "imageUrl" : "imageUrl",
          "name" : "name",
          "description" : "description",
          "id" : "id",
          "category" : "category"
        } ]
      }
    },
...
    "ResponseMessage" : {
      "type" : "object",
      "properties" : {
        "message" : {
          "type" : "string"
        }
      }
    },
...
    "Meal" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "category" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "price" : {
          "type" : "integer"
        },
        "imageUrl" : {
          "type" : "string"
        }
      },
      "example" : {
        "price" : 0,
        "imageUrl" : "imageUrl",
        "name" : "name",
        "description" : "description",
        "id" : "id",
        "category" : "category"
      }
    },
    "Menu" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "unixTime" : {
          "type" : "integer"
        },
        "expriry" : {
          "type" : "string"
        },
        "meals" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Meal"
          }
        }
      },
      "example" : {
        "expriry" : "expriry",
        "unixTime" : 0,
        "name" : "name",
        "id" : "id",
        "meals" : [ {
          "price" : 0,
          "imageUrl" : "imageUrl",
          "name" : "name",
          "description" : "description",
          "id" : "id",
          "category" : "category"
        }, {
          "price" : 0,
          "imageUrl" : "imageUrl",
          "name" : "name",
          "description" : "description",
          "id" : "id",
          "category" : "category"
        } ]
      }
    },
...
  }
}